to the method performIteration() was added check for very small digits in a table to change it fro the 0 to avoid errors

+++ sem4/OperationsResearch/SimplexMethod/SimplexMethod/LinearProgram.h
@@ -11,9 +11,9 @@ class LinearProgram
 private:
 	int variablesNumb; //The number of variables
 	int constraintsNumb; //The number of constraints
-	vector<int> objectiveFunc; //"c"- the vector with objective func coeff
-	vector<vector<int>>constraintsMatrix;// "A" - matrix of constraints coeff
-	vector<int>constraintsValues; //"b" - the vecor of constraints values
+	vector<double> objectiveFunc; //"c"- the vector with objective func coeff
+	vector<vector<double>>constraintsMatrix;// "A" - matrix of constraints coeff
+	vector<double>constraintsValues; //"b" - the vecor of constraints values
 
 	void isValidNumber(int& num, const std::string& message); // check the correnctnes of input data (included check for lettes and other symbols)
 
@@ -27,9 +27,9 @@ public:
 	//Getters 
 	int getVariablesNumb() const { return variablesNumb; }
 	int getConstraintsNumb() const { return constraintsNumb; }
-	vector<int> getObjectiveFunc() const { return objectiveFunc; }
-	vector<vector<int>> getConstraintsMatrix() const { return constraintsMatrix; }
-	vector<int> getConstraintsValues() const { return constraintsValues; }
+	vector<double> getObjectiveFunc() const { return objectiveFunc; }
+	vector<vector<double>> getConstraintsMatrix() const { return constraintsMatrix; }
+	vector<double> getConstraintsValues() const { return constraintsValues; }
 
 };
